feat: skip updating testing model if distro doesn't exist
This is particularly useful where distro can be defined in `unstable`
stage, for early testing/development purposes without introducing the
distro in, the more stable, `testing` stage.

When time comes, the stage model for the distro should be manually added
in `testing` folder and the process will pick that up from then on.